test(rs-platform-wallet/e2e): restore ID-007 — pin intentional not-monitored contract
ID-007 reverts the prior assertion-flip and once again pins the
**intentional** architecture: DIP-9 identity-authentication subfeature
addresses (subfeature 0..3, 6-component path
`m/9'/coinType'/5'/{0,1,2,3}'/identity_index'/key_index'`) are NOT in
`PlatformWalletInfo::monitored_addresses()`, sending Core duffs to one
does NOT increase the wallet's Core balance, and the UTXO set never
observes such a send.

Investigation rationale (documented in the file docstring and the
TEST_SPEC.md rationale block): dash-evo-tool — the canonical Platform
client — treats these addresses as pure key material;
`account_summary.rs:226-229` explicitly notes they "usually hold zero
balance"; `receive_address()` returns BIP-44 paths only; the UI hides
them outside developer-mode "Identity System" view. No standard flow
sends Layer-1 Dash to identity-auth addresses. The closed PR
`dashpay/rust-dashcore#554` was speculative future-proofing for a
hypothetical use case, not a fix for an active bug — its rejection was
correct.

Restoration shape:
- File renamed back from `id_007_identity_auth_addresses_monitored.rs`
  to `id_007_identity_auth_addresses_not_monitored.rs`; `cases/mod.rs`
  follows.
- Test fn renamed to `id_007_identity_auth_addresses_not_monitored`.
- `monitored_before` / `monitored_after` flipped from `contains` to
  `!contains` for both `auth_addr_zero` and `auth_addr_one`.
- `wait_for_core_balance` window restored to
  `CORE_BALANCE_NEGATIVE_WINDOW = 30s` and the call is asserted
  `.is_err()` (timeout EXPECTED).
- UTXO assertion flipped to `assert_eq!(utxo_count_to_auth_addr, 0)`.
- File docstring and `#[ignore]` reason rewritten to frame the test as
  a defensive pin of intentional behavior — green = architecture
  intact, red = regression to investigate before flipping.
- TEST_SPEC.md ID-007 entry rewritten in the same Pass/intentional
  framing; quick-index row updated; assertions/scenario/rationale
  blocks aligned.

Verification:
- `cargo fmt --all` — clean.
- `cargo check --tests --all-features -p platform-wallet` — clean.
- `cargo clippy --tests --all-features -p platform-wallet -- -D warnings`
  — clean.
- `cargo test --lib -p platform-wallet` — 141 passed, 0 failed.

End-to-end (`cargo test -- --ignored` on ID-007) is expected to PASS;
deferred to operator/Marvin verification — testnet not run from this
agent.
