+++ packages/rs-platform-wallet/tests/e2e/cases/mod.rs
@@ -12,7 +12,7 @@ pub mod id_001_register_identity_from_addresses;
 pub mod id_002_top_up_identity;
 pub mod id_003_identity_to_identity_transfer;
 pub mod id_005_identity_to_addresses_transfer;
-pub mod id_007_identity_auth_addresses_monitored;
+pub mod id_007_identity_auth_addresses_not_monitored;
 pub mod id_sweep_recovers_identity_credits;
 pub mod pa_001_multi_output;
 pub mod pa_001b_change_address_branch;
